chore: move `devices` from `Playwright` to `LocalUtils`
This avoids sending device on every remote connect.
